% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_confirmDeleteModal.R
\name{confirmDeleteModal}
\alias{confirmDeleteModal}
\title{confirmDeleteModal}
\usage{
confirmDeleteModal(ns, i)
}
\arguments{
\item{ns}{namespace}

\item{i}{dataset to confirm delete for}
}
\value{
shiny.tag
}
\description{
Launches a modal to check whether the user wants to delete
}
