% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_datasets_argon_table.R
\name{datasets_argon_table}
\alias{datasets_argon_table}
\title{datasets_argon_table}
\usage{
datasets_argon_table(ns)
}
\arguments{
\item{ns}{namespace}
}
\value{
shiny.tag
}
\description{
Creates an argon table to show datasets available in the app
}
