% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataViewer.R
\name{dataViewer}
\alias{dataViewer}
\title{an interactive addin to help visualise data and identifiy outliers}
\usage{
dataViewer(data, x, y, predictions = NULL, id_col = NULL, col = NULL,
  group = NULL, lm_fit = FALSE)
}
\arguments{
\item{data}{a data frame that contains the raw data}

\item{x}{the x variable}

\item{y}{the y variable}

\item{predictions}{a data frame that contains the predictions if desired}

\item{id_col}{the column that splits the data frame into separate plots}

\item{col}{an optional variable if different colours are desired on each plot}

\item{group}{an optional variable for grouping model predictions when there are multiple predictions on a single plot}

\item{lm_fit}{whether or not you want a linear model fit to be superimposed over the data. Defaults to FALSE}
}
\value{
a dataframe of points that were clicked
}
\description{
opens a pane from which you can select each set of data and select points to be dropped. The undo button gets rid of the last selection. Press "DONE" to get a dataframe of the selected points.
}
\details{
an interactive addin to help visualise data and identify outliers
}

