% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_delete_dataset.R
\name{delete_dataset}
\alias{delete_dataset}
\title{delete_dataset}
\usage{
delete_dataset(i)
}
\arguments{
\item{i}{index of the dataset to delete}
}
\description{
Deletes a dataset from dataset file and cache
}
