% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clicked.R
\name{get_clicked}
\alias{get_clicked}
\title{subsets the rows in the data that were selected in dataViewer}
\usage{
get_clicked(data, clicked_data)
}
\arguments{
\item{data}{the dataframe}

\item{clicked_data}{the dataframe of clicked points}
}
\description{
subsets the rows in the data that were selected in dataViewer
}

