% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unclicked.R
\name{get_unclicked}
\alias{get_unclicked}
\title{subsets the rows in the data that were not selected in dataViewer}
\usage{
get_unclicked(data, clicked_data)
}
\arguments{
\item{data}{the dataframe}

\item{clicked_data}{the dataframe of clicked points}
}
\description{
subsets the rows in the data that were not selected in dataViewer
}

