% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datasets_table.R
\name{mod_datasets_table_server}
\alias{mod_datasets_table_server}
\title{datasets_table Server Functions}
\usage{
mod_datasets_table_server(id)
}
\arguments{
\item{id}{id for instantiated module}
}
\description{
A Server module for uploading and viewing dataset
}
