% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datasets_table.R
\name{mod_datasets_table_ui}
\alias{mod_datasets_table_ui}
\title{datasets_table UI Function}
\usage{
mod_datasets_table_ui(id)
}
\arguments{
\item{id}{id for instantiated module}
}
\description{
A UI module for uploading and viewing dataset
}
