% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_dataset.R
\name{read_dataset}
\alias{read_dataset}
\title{read_dataset}
\usage{
read_dataset(i)
}
\arguments{
\item{i}{index of dataset to read}
}
\value{
data.frame
}
\description{
Reads in a dataset from the dataset list
}
