% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Kaggle Titanic dataset for demonstration purposes}
\format{
A data frame with 891 rows and 12 variables
}
\usage{
titanic
}
\description{
Kaggle Titanic dataset for demonstration purposes
}
\keyword{datasets}
