% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_update_data_cache.R
\name{update_data_cache}
\alias{update_data_cache}
\title{update_data_cache}
\usage{
update_data_cache(
  data_type,
  data_file,
  data_loc,
  data_url,
  data_name,
  cache,
  cache_dir
)
}
\arguments{
\item{data_type}{Type of data e.g. Tabular, Time-Series, Image, Text}

\item{data_file}{Data file object from fileInput()}

\item{data_loc}{Local or URL}

\item{data_url}{URL of the data to download}

\item{data_name}{Name of the dataset}

\item{cache}{Yes or No - whether to cache data, only valid for URL as local is automatically caches by fileInput()}

\item{cache_dir}{Directpry to cache the data - default = "./cache/cached_datasets"}
}
\value{
The return value, if any, from executing the function.
}
\description{
Takes input from dataModal and adds new dataset to app and cache
}
